refactor: ItemViewController 구현 및 ItemAddViewController 리팩토링
